fix(monitor): 修复容器监控缺少运行时长、IP地址、端口映射的问题
- 更新 ContainerInfo 结构体与 agent /containers/search 接口保持一致
- 新增 runTime, network[], ports[] 等字段
- 前端表格新增运行时长、IP地址、端口映射列
- 端口映射超过2个时显示 tooltip 并折叠

+++ core/app/api/v2/node.go
@@ -219,8 +219,8 @@ func (b *BaseApi) GetNodeContainers(c *gin.Context) {
 		for i := range containers {
 			if stats, ok := statsMap[containers[i].ContainerID]; ok {
 				containers[i].CPUPercent = stats.CPUPercent
-				containers[i].MemUsage = float64(stats.MemoryUsage)
-				containers[i].MemLimit = float64(stats.MemoryLimit)
+				containers[i].MemUsage = stats.MemoryUsage
+				containers[i].MemLimit = stats.MemoryLimit
 			}
 		}
 	}
